Fixes a 'bug' where inner classes would not be imported (always used FQCN). Here's what I know:
Back when codemodel was inside the JAXB RI, JAXB-166 was filed to report a bug where generated code would not compile. That was found to be a bug in the javac compiler, and BugParade 6431987 was filed for that. As a temporary fix, this line was added to the formatter to avoid generating bad code. The compiler bug has since been resolved (sometime in JDK6, which is EOL anyway these days), so it seems reasonable to go back to importing inner classes when possible. They will still be imported in full if there is a naming conflict with another imported type.