Enable reference caching for classes with are not visible to the classloader
Otherwise, when using such a class multiple times within the same
JClass, that very same class is not identified as the same when building
the model in the 'collecting phase'. That causes JFormatter to think
that those classes collide because having the same name, although being
the very same class. As a consequence, no import directive is created
for that class but referenced the fully qualified way instead.