Added support for super-wildcards.
Merged in an old patch from the CodeModel JIRA (https://java.net/jira/browse/CODEMODEL-11) to support things like generics like <? super A>.
Also changed some Javadoc to get it to build the archives on my system.
And a small change to the build.gradle to switch to a SNAPSHOT build and to get it to work without authenticator properties.